requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
pydantic==2.4.2
python-multipart==0.0.6
pylint==3.0.2
flake8==6.1.0
bandit==1.7.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
psycopg2-binary==2.9.9
boto3==1.28.64 